// ExistingCode/src/data-access/api/database/data-contracts.ts
export interface HoleStatus {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Description: string;
	HoleStatusId: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	holes: Hole[];
}
// ...
export interface DrillType {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Description: string;
	DrillTypeId: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	drillMethods: DrillMethod[];
	drillPlans: DrillPlan[];
}
// ...
export interface LeaseStatus {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Description: string;
	IsDefaultInd: boolean;
	LeaseStatusId: string;
	SortOrder: number;
	tenements: Tenement[];
}

export interface TenementStatus {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Description: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	TenementStatusId: string;
	tenements: Tenement[];
}

export interface DrillPatternCode {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Description: string;
	DrillPatternCodeId: string;
	ExampleFormat: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	TypicalPatternType?: string;
	TypicalSpacing_m?: number;
	drillPatterns: DrillPattern[];
}

export interface DrillPatternType {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Description: string;
	DrillPatternTypeId: string;
	GeometryType?: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	drillPatterns: DrillPattern[];
}

export interface DrillPattern {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	DrillPattern: string;
	Comments?: string;
	DrillPatternCode: string;
	DrillPatternId: string;
	DrillPatternType: string;
	DrillProgram: string;
	Organization: string;
	Orientation?: number;
	SpacingX?: number;
	SpacingY?: number;
	Target: string;
	drillPattern2: DrillPatternCode;
	drillPatternType: DrillPatternType;
	drillProgram: DrillProgram;
	organization: Organization;
	target: Target;
	drillPlans: DrillPlan[];
}

export interface SubTarget {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	SubTarget: string;
	Description?: string;
	SortOrder?: number;
	SubTargetId: string;
	Target: string;
	target: Target;
	collars: Collar[];
	drillPlans: DrillPlan[];
}

export interface Target {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Target: string;
	CentroidEasting?: number;
	CentroidNorthing?: number;
	Description?: string;
	Organization: string;
	Priority?: number;
	Project?: string;
	Radius_m?: number;
	SortOrder?: number;
	TargetCode?: string;
	TargetId: string;
	TargetType?: string;
	Tenement: string;
	organization: Organization;
	tenement: Tenement;
	collars: Collar[];
	drillPatterns: DrillPattern[];
	drillPlans: DrillPlan[];
	subTargets: SubTarget[];
}

export interface Tenement {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Tenement: string;
	AnnExpRent: number;
	CostCd: string;
	DataSource: string;
	/** @format date-time */
	EndDt?: string;
	/** @format date-time */
	GrantDt?: string;
	Holder: string;
	LeaseStatus: string;
	Organization: string;
	/** @format date-time */
	RegDt?: string;
	RegStatus: string;
	/** @format date-time */
	StartDt?: string;
	TenementId: string;
	costCd: CostCodes;
	holder: Company;
	leaseStatus: LeaseStatus;
	regStatus: TenementStatus;
	collars: Collar[];
	drillPlans: DrillPlan[];
	drillPrograms: DrillProgram[];
	sites: Site[];
	targets: Target[];
}
// ...
export interface RowStatus {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: number;
	Description: string;
	IsDefaultInd: boolean;
	RowStatusId: string;
	SortOrder: number;
	collars: Collar[];
	collarCoordinates: CollarCoordinate[];
	coreRecoveryRunLogs: CoreRecoveryRunLog[];
	cycloneCleanings: CycloneCleaning[];
	drillMethods: DrillMethod[];
	fractureCountLogs: FractureCountLog[];
	geologyCombinedLogs: GeologyCombinedLog[];
	magSusLogs: MagSusLog[];
	metaDataLogs: MetaDataLog[];
	ptSamples: PtSample[];
	ptSampleQcs: PtSampleQc[];
	rigSetups: RigSetup[];
	rockMechanicLogs: RockMechanicLog[];
	rockQualityDesignationLogs: RockQualityDesignationLog[];
	samples: Sample[];
	sampleQcs: SampleQc[];
	sampleRegisters: SampleRegister[];
	shearLogs: ShearLog[];
	specificGravityPtLogs: SpecificGravityPtLog[];
	standardSampleQcs: StandardSampleQc[];
	structureLogs: StructureLog[];
	structurePtLogs: StructurePtLog[];
	surveys: Survey[];
	surveyLogs: SurveyLog[];
	xrfSamples: XrfSample[];
	xrfSampleQcs: XrfSampleQc[];
}
// ...
export interface Organization {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Organization: string;
	Description?: string;
	OrganizationId: string;
	SortOrder?: number;
	collars: Collar[];
	coreRecoveryRunLogs: CoreRecoveryRunLog[];
	drillMethods: DrillMethod[];
	drillPatterns: DrillPattern[];
	drillPlans: DrillPlan[];
	drillPrograms: DrillProgram[];
	fractureCountLogs: FractureCountLog[];
	geologyCombinedLogs: GeologyCombinedLog[];
	holes: Hole[];
	metaDataLogs: MetaDataLog[];
	ptSamples: PtSample[];
	ptSampleQcs: PtSampleQc[];
	rigSetups: RigSetup[];
	rockMechanicLogs: RockMechanicLog[];
	rockQualityDesignationLogs: RockQualityDesignationLog[];
	samples: Sample[];
	sampleQcs: SampleQc[];
	shearLogs: ShearLog[];
	specificGravityPtLogs: SpecificGravityPtLog[];
	standardSamples: StandardSample[];
	standardSampleQcs: StandardSampleQc[];
	structureLogs: StructureLog[];
	surveys: Survey[];
	targets: Target[];
	xrfSamples: XrfSample[];
	xrfSampleQcs: XrfSampleQc[];
	xrfStandardSamples: XrfStandardSample[];
}
// ...
export interface Hole {
	ReportIncludeInd?: boolean;
	ValidationStatus?: HoleValidationStatusEnum;
	ValidationErrors?: string | null;
	RowStatus?: HoleRowStatusEnum;
	SupersededById?: string;
	ActiveInd?: boolean;
	rv?: string;
	/** @format date-time */
	CreatedOnDt?: string;
	CreatedBy?: string;
	/** @format date-time */
	ModifiedOnDt?: string;
	ModifiedBy?: string;
	HoleId: string;
	HoleStatus: string;
	Organization: string;
	holeStatus: HoleStatus;
	organization: Organization;
	holeNames: HoleName[];
}

export interface HolePurposeDetail {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Description: string;
	HolePurpose: string;
	HolePurposeDetailId: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	holePurpose: HolePurpose;
	drillPlans: DrillPlan[];
}

export interface HolePurpose {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Category?: string;
	Description: string;
	HolePurposeId: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	collars: Collar[];
	drillPlans: DrillPlan[];
	holePurposeDetails: HolePurposeDetail[];
}

export interface HoleType {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	CoreSize?: string;
	Description: string;
	HoleTypeId: string;
	IsDefaultInd: boolean;
	Method?: string;
	SortOrder: number;
	collars: Collar[];
	drillPlans: DrillPlan[];
}

export interface Phase {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Phase: string;
	Description?: string;
	PhaseId: string;
	Project: string;
	SortOrder?: number;
	collars: Collar[];
	drillPlans: DrillPlan[];
}

export interface Pit {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Pit: string;
	Description?: string;
	PitId: string;
	Project: string;
	SortOrder?: number;
	collars: Collar[];
	drillPlans: DrillPlan[];
}

export interface Prospect {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Prospect: string;
	Description?: string;
	Project: string;
	ProspectId: string;
	SortOrder?: number;
	project: Project;
	collars: Collar[];
	drillPlans: DrillPlan[];
}

export interface Zone {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Description: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	ZoneId: string;
	drillPlans: DrillPlan[];
}
// ...
export interface Grid {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Datum?: string;
	Description: string;
	EPSG?: number;
	GridId: string;
	IsDefaultInd: boolean;
	IsProjected: boolean;
	Proj4Text?: string;
	Projection?: string;
	SortOrder: number;
	SRID: number;
	Units?: string;
	collarCoordinates: CollarCoordinate[];
	drillPlans: DrillPlan[];
	samples: Sample[];
	siteCoordinates: SiteCoordinate[];
	surveyLogs: SurveyLog[];
}
// ...
export interface Project {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Project: string;
	Description?: string;
	Organization: string;
	ProjectId: string;
	SortOrder?: number;
	collars: Collar[];
	drillPlans: DrillPlan[];
	drillPrograms: DrillProgram[];
	filteredsets: Filteredset[];
	prospects: Prospect[];
}
// ...
export interface VwDrillPlan {
	ActiveInd: boolean;
	CreatedBy: string;
	/** @format date-time */
	CreatedOnDt: string;
	DataSource: string;
	DrillHoleId: string;
	DrillPattern?: string;
	DrillPlanId: string;
	DrillPriority: number;
	DrillType?: string;
	Grid?: string;
	HoleNm?: string;
	HolePurpose?: string;
	HolePurposeDetail?: string;
	HoleStatus?: string;
	HoleType?: string;
	InfillTarget?: string;
	ModifiedBy?: string;
	/** @format date-time */
	ModifiedOnDt?: string;
	ODSPriority: number;
	Organization: string;
	OtherHoleNm?: string;
	Phase?: string;
	Pit?: string;
	PlannedAzimuth?: number;
	PlannedBy?: string;
	/** @format date-time */
	PlannedCompleteDt?: string;
	PlannedDip?: number;
	PlannedEasting?: number;
	PlannedHoleNm?: string;
	PlannedNorthing?: number;
	PlannedRL?: number;
	/** @format date-time */
	PlannedStartDt?: string;
	PlannedTotalDepth?: number;
	Priority?: number;
	Project: string;
	ProposedHoleNm?: string;
	Prospect?: string;
	QCInsertionRuleId?: string;
	ReportIncludeInd: boolean;
	SitePrep?: string;
	SubTarget?: string;
	SupersededById?: string;
	Target?: string;
	Tenement?: string;
	TWF?: string;
	ValidationErrors?: string;
	ValidationStatus: number;
	vwDrillPlanId: string;
	WaterTableDepth?: number;
	Zone?: string;
	RowStatus?: VwDrillPlanRowStatusEnum;
	rv?: string;
}
// ...
export type VwDrillPlanRowStatusEnum = 0 | 1 | 2 | 3 | 4;